pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-tavily>=0.2.10",
    "langgraph>=0.5.3",
    "langgraph-cli[inmem]>=0.3.5",
    "langsmith>=0.4.7",
    "mcp>=1.12.0",
    "python-dotenv>=1.1.1",
]
